pyproject.toml
[project]
name = "gemini-ocr-tool"
version = "0.1.0"
description = "A Python CLI tool"
requires-python = ">=3.14"
dependencies = [
    "click>=8.1.7",
    "google-genai>=1.0.0",
    "pillow>=12.0.0",
    "tqdm>=4.66.0",
]
authors = [
    {name = "Dennis Vriend", email = "dvriend@ilionx.com"}
]
license = {text = "MIT"}
readme = "README.md"
keywords = ["cli", "tool"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
]

[project.urls]
Homepage = "https://github.com/dnvriend/gemini-ocr-tool"
Repository = "https://github.com/dnvriend/gemini-ocr-tool"
Issues = "https://github.com/dnvriend/gemini-ocr-tool/issues"

[project.scripts]
gemini-ocr-tool = "gemini_ocr_tool.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "bandit>=1.7.0",
    "pip-audit>=2.6.0",
    "types-tqdm>=4.66.0",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
python_version = "3.14"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
strict = true

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # Skip assert_used (common in tests)
